ai_callbot_crew/requirements.txt
crewai
numpy>=1.26.0 ;
SpeechRecognition>=3.10.0
pyaudio>.env/=0.2.14 ;
coqui-tts
pydantic

# Text-to-Speech (Coqui TTS) — installs Torch as a dependency
coqui-tts

# Misc utilities
pysbd>=0.3.4